% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_dyn_available}
\alias{is_dyn_available}
\title{Check if mecab or its dynamic libarary is available}
\usage{
is_dyn_available()
}
\value{
Logical.
}
\description{
Check if mecab or its dynamic libarary is available
}
\keyword{internal}
