% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isBlank}
\alias{isBlank}
\title{Check if scalars are blank}
\usage{
isBlank(x, trim = TRUE, ...)
}
\arguments{
\item{x}{Object to check its emptiness.}

\item{trim}{Logical.}

\item{...}{Additional arguments for \code{base::sapply()}.}
}
\value{
Logical values.
}
\description{
Check if scalars are blank
}
