% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isDynAvailable}
\alias{isDynAvailable}
\title{Check if dynamic libraries are available}
\usage{
isDynAvailable(dynlib = "libmecab")
}
\arguments{
\item{dynlib}{The names of dynamic libraries to check their accessibility. Default value is `libmecab`.}
}
\value{
Logical.
}
\description{
Check if dynamic libraries are available
}
