% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pack}
\alias{pack}
\title{Pack Output of POS Tagger}
\usage{
pack(df, collapse = " ")
}
\arguments{
\item{df}{Output of \code{pos(format = "data.frame")}, \code{posParallel(format = "data.frame")} or \code{posSimple}.}

\item{collapse}{This argument will be passed to \code{paste()}.}
}
\value{
data.frame
}
\description{
Pack Output of POS Tagger
}
\examples{
\dontrun{
sentence <- c("some UTF-8 texts")
result <- pos(sentence, format = "data.frame")
pack(result)
}

}
