% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posApplyDFRcpp}
\alias{posApplyDFRcpp}
\title{Call POS Tagger via `lapply` and return a list of data.frame}
\arguments{
\item{text}{Character vector.}

\item{sys_dic}{String scalar.}

\item{user_dic}{String scalar.}
}
\value{
list of data.frame.
}
\description{
Call POS Tagger via `lapply` and return a list of data.frame
}
\keyword{internal}
