% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posApplyJoinRcpp}
\alias{posApplyJoinRcpp}
\title{Call POS Tagger via `lapply` and return a named list.}
\arguments{
\item{text}{Character vector.}

\item{sys_dic}{String scalar.}

\item{user_dic}{String scalar.}
}
\value{
named list.
}
\description{
Call POS Tagger via `lapply` and return a named list.
}
\keyword{internal}
