% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posLoopDFRcpp}
\alias{posLoopDFRcpp}
\title{Call tagger inside a loop and return a data.frame}
\arguments{
\item{text}{Character vector.}

\item{sys_dic}{String scalar.}

\item{user_dic}{String scalar.}
}
\value{
data.frame.
}
\description{
Call tagger inside a loop and return a data.frame
}
\keyword{internal}
