% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posLoopRcpp}
\alias{posLoopRcpp}
\title{Call tagger inside a loop and return a list of named character vectors.}
\arguments{
\item{text}{Character vector.}

\item{sys_dic}{String scalar.}

\item{user_dic}{String scalar.}
}
\value{
list of named character vectors.
}
\description{
Call tagger inside a loop and return a list of named character vectors.
}
\keyword{internal}
