% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posSimple.R
\name{posSimple}
\alias{posSimple}
\title{Yet another part-of-speech tagger}
\usage{
posSimple(
  sentence,
  into = c("POS1", "POS2", "POS3", "POS4", "X5StageUse1", "X5StageUse2", "Original",
    "Yomi1", "Yomi2"),
  sys_dic = "",
  user_dic = ""
)
}
\arguments{
\item{sentence}{Character vector.}

\item{into}{Character vector.}

\item{sys_dic}{A location of system MeCab dictionary. The default value is "".}

\item{user_dic}{A location of user-specific MeCab dictionary. The default value is "".}
}
\value{
data.frame
}
\description{
Yet another part-of-speech tagger
}
