% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppMeCab-package.r
\docType{package}
\name{RcppMeCab}
\alias{RcppMeCab}
\alias{RcppMeCab-package}
\title{RcppMeCab: Rcpp Wrapper for MeCab Library}
\description{
R package based on 'Rcpp' for 'MeCab': Yet Another Part-of-Speech and
Morphological Analyzer (\url{http://taku910.github.io/mecab/}). The purpose of
this package is providing a seamless developing and analyzing environment for
CJK texts. This package utilizes parallel programming for providing
highly efficient text preprocessing \code{posParallel()} function.
For installation, please refer to \href{https://github.com/junhewk/RcppMeCab/blob/master/README.md}{README.md} file.
}
\details{
This package utilizes 'MeCab' C API and 'Rcpp' codes.
}
\references{
\itemize{
\item{\href{http://taku910.github.io/mecab/}{MeCab}}
\item{\href{http://www.rcpp.org}{Rcpp: Seamless R and C++ Integration}}
\item{\href{http://eunjeon.blogspot.com}{Eunjeon project}}
}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/paithiov909/RcppMeCab/issues}
}

}
\author{
Junhewk Kim
}
\keyword{Chinese}
\keyword{Japanese}
\keyword{Korean}
\keyword{MeCab}
\keyword{morpheme}
\keyword{nlp}
\keyword{part-of-speech}
