% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isDynAvailable}
\alias{isDynAvailable}
\title{Check if mecab or its dynamic libarary is available}
\usage{
isDynAvailable()
}
\value{
Logical.
}
\description{
Check if mecab or its dynamic libarary is available
}
