% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pack}
\alias{pack}
\title{Pack Output of POS Tagger}
\usage{
pack(df, pull = "token", .collapse = " ")
}
\arguments{
\item{df}{Output of \code{pos(format = "data.frame")} or \code{posParallel(format = "data.frame")}.}

\item{pull}{Column name to be packed into data.frame. Default value is `token`.}

\item{.collapse}{This argument will be passed to \code{stringr::str_c()}.}
}
\value{
data.frame
}
\description{
Pack Output of POS Tagger
}
\examples{
\dontrun{
sentence <- c("some UTF-8 texts")
result <- pos(sentence, format = "data.frame")
pack(result)
}

}
