% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posSimple.R
\name{posSimple}
\alias{posSimple}
\title{yet another part-of-speech tagger}
\usage{
posSimple(sentence, into = NULL, sys_dic = "", user_dic = "")
}
\arguments{
\item{sentence}{Character vector.}

\item{into}{NULL or character vector. If not NULL, this argument is passed to \code{tidyr::separate}}

\item{sys_dic}{A location of system MeCab dictionary. The default value is "".}

\item{user_dic}{A location of user-specific MeCab dictionary. The default value is "".}
}
\value{
data.frame
}
\description{
\code{posSimple} returns part-of-speech (POS) tagged morphemes of the sentence as a simple data.frame.
}
\details{
This function returns morphemes and their features as a simple data.frame.
If only a character vector provided as `into`, the function separates every feature with `,`
into the dictionary-specific format.

It is useful when you want to check all features of morphemes and then filter, remove or count
some morphemes by their features that are not available in output of \code{pos} and \code{posParallel}.
}
\examples{
\dontrun{
sentence <- c("some UTF-8 texts")
posSimple(sentence)
# Parse features of Japanese IPA dictionary.
posSimple(sentence, into = c("POS1", "POS2", "POS3", "POS4", "X5StageUse1", "X5StageUse2", "Original", "Yomi1", "Yomi2"))
}
}
